pyproject.toml
[project]
name = "marketing-automation-mcp"
version = "1.1.0"
description = "A Model Context Protocol server for marketing campaign automation workflows."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Mohit4022-cloud" }
]
dependencies = [
  "mcp==1.27.0",
  "httpx>=0.27.0",
  "google-auth>=2.28.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "openai>=2.31.0",
  "anthropic>=0.94.1",
  "google-genai>=1.72.0",
  "jinja2>=3.1.3",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "pydantic>=2.13.0",
  "pydantic-settings>=2.2.0",
  "email-validator>=2.1.0",
  "apscheduler>=3.10.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "plotly>=5.18.0",
  "kaleido>=0.2.1",
  "weasyprint>=60.0",
  "structlog>=24.0.0",
  "python-json-logger>=2.0.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "python-dateutil>=2.9.0",
  "jsonschema>=4.21.0",
  "click>=8.1.0",
  "tabulate>=0.9.0",
  "cryptography>=42.0.0",
  "keyring>=24.0.0",
  "pyjwt>=2.8.0",
  "psutil>=5.9.0",
  "aioredis>=2.0.0",
  "flask>=3.0.0",
  "psycopg2-binary>=2.9.9"
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0",
  "flake8>=7.0.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "faker>=22.0.0",
  "responses>=0.24.0"
]

[project.scripts]
marketing-automation = "src.cli:cli"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
^/migrations/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", "venv", ".venv", "*.egg-info"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["src", "src.*"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
